client line buffering
Perform line-buffering by default. The client content is not written to
the backend until a newline is detected. This allows merging of separate
client writes to entire lines.

The global option `no-line-buffering` disables the line buffered mode.
The per-client option `line-buffering` enables this option per client if
the global setting is disabled.

After a certain amount of time, the client output is written even if no
newline was detected. The timeout can be controlled globally and
per-client using the `line-buffering-ms` option. The default timeout is
50ms giving a moderate terminal experience.

See ticket #CD-1061.

Co-authored-by: Georg Kotheimer <georg.kotheimer@kernkonzept.com>
Change-Id: I74f58ac18dfe054d4f8e681505597b5b0f1d6f64