make vcon_read signal end of data when input buffer is exactly filled
L4::Vcon.read() specifies that more bytes to read are available only
if the number of bytes read exceeds the buffer size passed.
On equality of buffer size and bytes read, there is no more data
available.
Therefore, vcon_read needs to set the L4_VCON_READ_STAT_DONE flag on
equality of both values as well, to signal completion to the client.

Change-Id: I0739b1afcc81ad86b81ed1a05fa0879359a0e00c