vcon_read: fix more to read signaling; add Buf.empty() function
vcon_read did not signal when there is more data to be read. In this
case the specification requires to return a value greater than the
input buffer size.
vcon_read signaled more data only by not setting the READ_STATS_DONE
flag, which is insufficient.

With a recent change in L4Re::Util::Vcon_svr this issue was
highlighted.

Furthermore, this change improves readbility of the affacted code
by introducing a dedicated empty() function to Client::Buf.

Change-Id: Id830dd875a36fbc78cd169a3783822ebdc265cd3