Revert "make vcon_read signal end of data when input buffer is exactly filled"
This reverts commit 8f26f9397a44e03a2cd8f091e2b3864bdecd4305.

Reason for revert: The changed check is the companion check for l.26.
If `r < size` then we know all is read.
If `r < size` is false, we either have `r == size` or `r > size`. 
If `r > size`, rbuf()->distance() should be greater zero anyways.
If `r == size`, rbuf()->distance() is greater than zero and there is more to read or rbuf()->distance() is zero and we have exactly read everything.
If we have more to read we have set `r = size` in line 27.

The check for `r <= size` is flawed, because we lose the case where we executed l.27. With the check for equality the second part of l.46 is not executed anymore.

Change-Id: Ib6208ba6ca73bdfc34c4b4518c12a7377f3c9d78