on "connect" and "show", drop unseen buffer content of this client
When connecting to a client which had its output hidden before, the last
10 lines of the client buffer are shown. However, when the client now
produces more output, the entire so far unseen client buffer is shown.
This behavior is unintended, because the last 10 lines of the client
were already shown.

There is a similar behavior with "show". Performing this command intends
to show the client output from now on. If the client now produces more
output, the entire so far unseen client buffer is shown. This behavior
is unintended as well.

Fix this by explicitly mark the client buffer as "seen" in both cases.

Change-Id: I72b5d271cb922541b5e82ae4dbbda1b1547a879c