Deregister client from controller on destruction
The controller keeps a list of pointers to clients. If a client gets
deleted because the corresponding IPC gate is deleted, the pointer to
that deleted client stays in the controller's list. The controller will
still find that client via its tag and connect to it if the user
requests it, but then a segfault happens.

To fix this, each client now gets a pointer to the controller it's
registered with, if any. On destruction of the client, it removes itself
from the controller's list.

Change-Id: I445afc33871e3c8748c5fc4a0c18f01e7fc97ed8