fix client index assignment in case of duplicated client tags
When a new client is created at cons, the client tag is checked against
all existing clients. If another client with the same tag is found,
the new client get's the found client's index plus one.
Then, the new client get's added to the collection of clients.

Before I58ed4f556aca57a5117f7b46358bb05ee4ccd147 the collection type
was a list and the new client was inserted before the old client,
ensuring the client with the highest index is found first.
The change changed the collection type vector and the search algorithm
remained the same. However, a vector doesn't offer means to insert
before a specific element and instead the new client get's pushed to the
end of the vector.
Thus, the first client found is always the same first client with this
tag and all subsequent clients with the same tag get 'tag:1' as
identifier.

When the user reboots clients and gets multiple clients with the same
'tag:1' tags, it can only select the first 'tag:1' client. All other
remain inaccessible.

This change fixes the index assignment, such that clients again get the
intended unique `tag:N` tag.

Change-Id: I7ac7f623b415e62d017a5bc610439e33b210eccf