fix second part of client index assignment
I7ac7f623b415e62d017a5bc610439e33b210eccf improved the counting,
however, in case the user drops a client, the index is duplicated
again.
The original behavior was:
1. forward scan the client list for a client with duplicate tag,
2. assign the new client the found client's index plus 1,
3. insert the new client before the found client.

Since the client list is a vector now and we only add elements to the
back of the vector, we do the above with a reverse search to
reinstate the original behavior.
1. reverse scan the client list for aclient with duplicated tab,
2. assign the new client the found client's index plus 1,
3. inser the new client after the found client.

Change-Id: I81c188273dd86dcbcc149f3a0d06319295dac639