Fix contradictory return value in Vcon_client::vcon_read
Buf::get() returns a byte array until either the head pointer or
the end of the underlying ring buffer, whatever comes first.
In case the ring buffer wrapped, the wrapped around part is not returned
and must be queried for again.

The return value of get() is the length of this byte array.
When this value is longer than the caller's buffer, the value gets
reduced to the size of the caller's buffer.
Then the copying happens and the number of copied bytes is cleared
from the ring buffer. If the copying encounters a break signal pending
on the byte to be copied, the copying stops.

Then there is the faulty check. If all characters that should be copied
are copied (no break signal encountered) and the bytes in the buffer
are less than the input buffer's size, then READ_STATS_DONE is set.
Which is false, because the buffer is not empty, due to the wrapped
around part in the buffer.
The reason for the logical OR included in this check is unknown.

Since Change-Id: Id830dd875a36fbc78cd169a3783822ebdc265cd3 another
check for the ring buffer being empty follows. In case the ring buffer
is not empty, the return value is set to 'size + 1', signaling more
data to read.
Now we have a contradictory return value combining READ_STATS_DONE with
'more to read'.

In case of a break signal detection during copy, we should return only
the number of copied bytes.
If no break signal is detected, the return value depends on more data
being available to read, which either means returning the input buffer
size + 1 or returning READ_STAT_DONE together with the number of bytes
read.

Change-Id: Ie4d3b2ce3cde9dcb174509d8a36774967f353866
Signed-off-by: Philipp Eppelt <philipp.eppelt@kernkonzept.com>