fix break signal detection in vcon_client
The break signal shall signal the UART client the start of a new
command sequence. Cons buffers this signal when sent from hardware.

When cons encounteres the signal while already reading characters from
the buffer, it stops reading characters and delivers all read characters
up to the signal to the client.
The next client read of said buffer will then start at the break signal
and return the signal together with the character buffer.

Up to now the processing of the break signal was flawed, because break
signals were checked for at the wrong end of the buffer. This could lead
to break signals not being detected during read.

This adds function documentation for the break signal handling as well.

Change-Id: I63a372a09b6f5e1310e73f42f24ab139bf42aea2
Signed-off-by: Philipp Eppelt <philipp.eppelt@kernkonzept.com>