vcon_read: read maximum number of available characters
The client specifies a number of bytes it wants to read. In case the
underlying ring-buffer wrapped around, vcon_read() read only until
the end of the buffer, even if there number of bytes the client wants
to read wasn't reached.

The new code handles the wrap-around-case and reads until either no
more data is available or the client's read is fully satisfied.

Change-Id: I3b555f733ae232445adc2ac84b8cfc02e92618a9
Signed-off-by: Philipp Eppelt <philipp.eppelt@kernkonzept.com>