prevent accidental out-of-bounds access of heap memory
Buf::get() returns a char const * to any place within the buffer and
the number of bytes readable starting from this location. The caller
must ensure to not read more than the returned number of bytes.

Use of the returned char const * as C-style string is not allowed,
because the char const * is not zero-terminated. strlen() et al will not
work properly.
Instead they either read stale content from the buffer, or worst case
read beyond the buffer and read other data on the heap.

To prevent any C-style string reads beyond the buffer, we allocate an
additional byte and set it to zero. This byte is not used in normal
ring buffer operation and remains in place as safeguard.

Change-Id: I835d75283ed19a7e7c1d26d724b0cc43f55b4d9c
Signed-off-by: Philipp Eppelt <philipp.eppelt@kernkonzept.com>